chore(gatsby): move common-topics to detail template, fetch data with GQL
