+++ gatsby/src/pages/index.tsx
@@ -2,8 +2,6 @@ import React from 'react';
 import { Helmet } from 'react-helmet';
 import { graphql } from 'gatsby';
 import ContentBox from '@/components/content-box';
-import CommonTopic from '@/components/common-topic';
-import LinkList from '@/components/link-list';
 import { Link } from 'gatsby-plugin-react-i18next';
 import MeasureList from '@/components/measure-list';
 import Container from '@/components/container';
@@ -26,37 +24,6 @@ const Home: React.FC<IProps> = ({ data }) => {
   } = homepage;
   const { measure_items, situation_items } = relationships;
 
-  const mockLinks = [
-    {
-      title: 'Autoservisy a pneuservisy',
-      href: '#0',
-    },
-    {
-      title: 'Mytí a čištění aut',
-      href: '#0',
-    },
-    {
-      title: 'Čerpací stanice',
-      href: '#0',
-    },
-    {
-      title: 'Autoservisy a pneuservisy 2',
-      href: '#0',
-    },
-    {
-      title: 'Mytí a čištění aut 2',
-      href: '#0',
-    },
-    {
-      title: 'Čerpací stanice 2',
-      href: '#0',
-    },
-    {
-      title: 'Čerpací stanice 3',
-      href: '#0',
-    },
-  ];
-
   return (
     <>
       <Helmet title="Covid Portál" />
@@ -78,9 +45,6 @@ const Home: React.FC<IProps> = ({ data }) => {
         >
           <MeasureList measures={measure_items} />
         </ContentBox>
-        <CommonTopic title="Podobná témata" boldedTitleCount={1}>
-          <LinkList links={mockLinks} variant="secondary" />
-        </CommonTopic>
         <Link to="/" language="en" className="">
           English Page
         </Link>
